Added some styling to taskboardbuttons.

+++ apps/client/src/components/Card/TaskboardCard.svelte
@@ -1,13 +1,19 @@
 <script lang="ts">
 	import { createEventDispatcher } from "svelte";
 	import BaseCard from "./BaseCard.svelte";
-	import type { ClientTaskboard } from "data-transfer-interfaces/taskboard/clientTaskboard";
+
 	import { TaskboardService } from "../../services/taskboardService";
+	import type { ClientTaskboard } from "data-transfer-interfaces/taskboard/clientTaskboard";
 
 	const dispatch = createEventDispatcher();
 	export let taskboard: ClientTaskboard | null;
+	let menuToggle: boolean = false;
 
 	const deleteTaskboard = async () => {
+		menuToggle = false;
+		const conf = confirm(`Are you sure you want to delete ${taskboard.name}?`);
+		if (!conf) return;
+
 		const res = await TaskboardService.deleteTaskboard(taskboard.uri);
 		if (res.ok) {
 			dispatch("deletedTaskboard", res.data);
@@ -18,18 +24,43 @@
 </script>
 
 <BaseCard>
-	<div class="grid grid-rows-3 h-full w-full">
-		<div>
-			<p class="">{taskboard.name}</p>
+	<div
+		class="h-full w-full flex flex-col transition-colors hover:bg-neutral-600"
+	>
+		<div class="flex-1">
+			<button class="block w-full h-28">
+				<p>{taskboard.name}</p>
+			</button>
 		</div>
 
-		<div />
+		<div
+			class={`flex-none flex ${
+				!menuToggle && "flex-row-reverse pointer-events-none"
+			} -mt-8 h-8`}
+		>
+			<div hidden={!menuToggle} class="flex-1">
+				<button
+					on:click={deleteTaskboard}
+					class="h-full w-full hover:text-red-900 hover:bg-white">delete</button
+				>
+			</div>
 
-		<div>
-			<button
-				class="mt-1 mr-0.5 float-right text-red-500 hover:text-red-700 active:text-black h-6 w-6"
-				on:click={deleteTaskboard}>X</button
-			>
+			<div hidden={!menuToggle} class="flex-1" />
+
+			<div hidden={!menuToggle} class="flex-1" />
+
+			<div hidden={!menuToggle} class="flex-1" />
+
+			<div class="flex-none pointer-events-auto">
+				<button
+					on:click={() => (menuToggle = !menuToggle)}
+					class={`block float-right p-1 h-8 w-8 rotate-90 ${
+						menuToggle ? "rounded-none" : "rounded-l-full rounded-br-full"
+					} hover:bg-neutral-400 active:bg-neutral-500`}
+				>
+					<img src="/dots.svg" alt="dots" />
+				</button>
+			</div>
 		</div>
 	</div>
 </BaseCard>
